Finish the report's skeleton
I am unbelievably close to death right now.

Many bugs were fixed, and many .pngs were added.

Task 13 and 14 remain unsolved.

+++ L1/Z1_8.cpp
@@ -9,7 +9,7 @@
 
 int main() {
     unsigned short int E, S, N;
-    long int tmp; /* in case the user _really_ wants this to break */
+    int tmp;
 
     p_fix_locale();
